Extractive Text Summarization VS Abstractive Text Summarization
Text summarization is the process of summarizing huge texts into smaller ones without changing the actual meaning of the sentences. There are 2 types of text summarization, namely extractive and abstractive:

Extractive text summarization is the process of summarizing the data where no new words are generated in the summary, i.e the entire summary is made up of words present in the original text.

Abstractive text summarization is the process of summarizing the data where new words may or mayn't be generated in the summary, i.e the summary may contain words other than the ones present in the original text.

In this project we will work with abstract text summarization.